requirements.txt
phidata
fastapi
tavily-python
crawl4ai
pydantic
web3
dotenv
typing